requirements.txt
bibtexparser==1.4.0
click==8.1.3
click_option_group==0.5.3
numpy==1.22.0
PyYAML==6.0
Requests==2.31.0
wikibaseintegrator==0.12.3
